feat: adopt RESTful API style and unify code structure

+++ src/main/java/com/loyce/omniflow/dao/entity/NodeClosureDO.java
@@ -9,12 +9,12 @@ public class NodeClosureDO {
     /**
      * 祖先节点id
      */
-    private Integer ancestor;
+    private Long ancestor;
 
     /**
      * 后代节点id
      */
-    private Integer descendant;
+    private Long descendant;
 
     /**
      * 层级深度 0表示自身
@@ -24,5 +24,5 @@ public class NodeClosureDO {
     /**
      * 所属库id
      */
-    private Integer libraryId;
+    private Long libraryId;
 }
