+++ src/main/java/com/loyce/omniflow/dto/req/NodeMoveReqDTO.java
@@ -13,16 +13,16 @@ public class NodeMoveReqDTO {
     /**
      * 要移动的节点 ID
      */
-    private Integer nodeId;
+    private Long nodeId;
 
     /**
      * 新的父节点 ID
      */
-    private Integer newParentId;
+    private Long newParentId;
 
     /**
      * 库 ID
      */
-    private Integer libraryId;
+    private Long libraryId;
 
 }
