+++ src/main/java/com/loyce/omniflow/service/helper/NodeNameConflictChecker.java
@@ -19,7 +19,7 @@ public class NodeNameConflictChecker {
      * @param libraryId 所属库 ID
      * @param excludeId 要排除的节点 ID（如重命名或移动时排除自己），可为 null
      */
-    public void checkDuplicateName(String name, Integer parentId, Integer libraryId, Integer excludeId) {
+    public void checkDuplicateName(String name, Long parentId, Long libraryId, Long excludeId) {
         int count = nodeMapper.countByNameAndParent(name, parentId, libraryId, excludeId);
         if (count > 0) {
             throw new ClientException("同级目录下已存在同名节点：" + name);
